Use hmatrix library
Benchmark 3D stayed at 0.06s, benchmark 4D improved slightly (0.36s
to 0.31s). I'm pretty sure it's possible to do way better than that,
but the main advantage is having access to all the pre-written linear
algebra stuff. I'm not sure how much is relevant, but at least the
determinant and inverse are.

The library is buggy, which is not ideal, but it had the best
combination of features of any linear algebra library I looked at,
and it seems the bugs can be worked around.