Add more steps towards generating the possible symmetries.
The feature is in an incomplete state but I want to try out variations
without losing what I have. It can filter the list of symmetries in
a given dimension to a much smaller list of unique ones under an
incomplete but cheap set of transformations. There's an attempt to
get it to then generate the list of point groups from this, but
that's incomplete because I'm having trouble with working out which
symmetries are compatible with the same norms.